Add Marshal.PtrToStringUni internal implementation in S.P.CoreLib
This change adds and internal Marshal.PtrToStringUni in S.P.CoreLib to unblock  the benchmark Sergiy has been working on.

[tfs-changeset: 1662877]
