+++ src/System.Private.CoreLib/src/Interop/Interop.manual.cs
@@ -129,51 +129,3 @@ internal partial class Interop
             uint dwFlags);
     }
 }
-
-namespace System.Runtime.InteropServices
-{
-    internal class Marshal
-    {
-        public static int GetLastWin32Error()
-        {
-            return PInvokeMarshal.GetLastWin32Error();
-        }
-
-        public static unsafe IntPtr AllocHGlobal(IntPtr cb)
-        {
-            return PInvokeMarshal.AllocHGlobal(cb);
-        }
-
-        public static unsafe IntPtr AllocHGlobal(int cb)
-        {
-            return PInvokeMarshal.AllocHGlobal(cb);
-        }
-
-        public static void FreeHGlobal(IntPtr hglobal)
-        {
-            PInvokeMarshal.FreeHGlobal(hglobal);
-        }
-
-        public static unsafe IntPtr AllocCoTaskMem(int cb)
-        {
-            return PInvokeMarshal.AllocCoTaskMem(cb);
-        }
-
-        public static void FreeCoTaskMem(IntPtr ptr)
-        {
-            PInvokeMarshal.FreeCoTaskMem(ptr);
-        }
-
-        public static void Copy(IntPtr source, byte[] destination, int startIndex, int length)
-        {
-            InteropExtensions.CopyToManaged(source, destination, startIndex, length);
-        }
-
-#if PLATFORM_UNIX
-        public static unsafe String PtrToStringAnsi(IntPtr ptr)
-        {
-            return PInvokeMarshal.PtrToStringAnsi(ptr);
-        }
-#endif
-    }
-}
